+++ hammurabi/__init__.py
@@ -26,6 +26,7 @@ from hammurabi.rules.ini import (
     SectionRenamed,
 )
 from hammurabi.rules.operations import Copied, Moved, Renamed
+from hammurabi.rules.templates import TemplateRendered
 from hammurabi.rules.text import LineExists, LineNotExists, LineReplaced
 
 __version__ = "0.1.2"